+++ apps/web/app/(entity)/[network]/@rightpanel/[...path]/page.tsx
