+++ apps/web/app/(entity)/[network]/@rightpanel/default.tsx
