+++ apps/web/app/(entity)/[network]/[...path]/layout.tsx
